+++ app/sidekiq/extraction_job.rb
@@ -15,7 +15,7 @@ class ExtractionJob
     @job.mark_as_running
     @job.update(start_time: Time.zone.now)
 
-    ExtractionExecution.new(@job, @job.extraction_definition).call
+    Extraction::Execution.new(@job, @job.extraction_definition).call
 
     @job.mark_as_completed unless @job.cancelled?
     @job.update(end_time: Time.zone.now)
